+++ realtime/realtime.js
@@ -110,31 +110,28 @@ export class Realtime {
         this.staging = staging; 
         this.opts = opts;
 
-        var staginURL = process.env.PROXY ? [
-                "tls://api2.relay-x.io:8666"
-                ] : [
-                    "nats://0.0.0.0:4221",
-                    "nats://0.0.0.0:4222",
-                    "nats://0.0.0.0:4223",
-                ]
-
-        if (staging !== undefined || staging !== null){
-            this.#baseUrl = staging ? staginURL : 
-                [
+        if(process.env.PROXY){
+            this.#baseUrl = ["tls://api2.relay-x.io:8666"];
+            initDNSSpoof();
+        }else{
+            if (staging !== undefined || staging !== null){
+                this.#baseUrl = staging ? [
+                        "nats://0.0.0.0:4221",
+                        "nats://0.0.0.0:4222",
+                        "nats://0.0.0.0:4223",
+                    ] : 
+                    [
+                        `tls://api2.relay-x.io:4221`,
+                        `tls://api2.relay-x.io:4222`,
+                        `tls://api2.relay-x.io:4223`
+                    ];
+            }else{
+                this.#baseUrl = [
                     `tls://api.relay-x.io:4221`,
                     `tls://api.relay-x.io:4222`,
                     `tls://api.relay-x.io:4223`
                 ];
-        }else{
-            this.#baseUrl = [
-                `tls://api.relay-x.io:4221`,
-                `tls://api.relay-x.io:4222`,
-                `tls://api.relay-x.io:4223`
-            ];
-        }
-
-        if(process.env.PROXY){
-            initDNSSpoof();
+            }
         }
 
         this.#log(this.#baseUrl);
@@ -746,7 +743,9 @@ export class Realtime {
             var arrayCheck = ![CONNECTED, DISCONNECTED, RECONNECT, this.#RECONNECTED,
                 this.#RECONNECTING, this.#RECONN_FAIL, MESSAGE_RESEND, SERVER_DISCONNECT].includes(topic);
 
-            var spaceStarCheck = !topic.includes(" ") && !topic.includes("*") && !topic.includes(".");
+            const TOPIC_REGEX = /^(?!\$)[A-Za-z0-9_,.*>\$-]+$/;
+
+            var spaceStarCheck = !topic.includes(" ") && TOPIC_REGEX.test(topic);
 
             return arrayCheck && spaceStarCheck;
         }else{