+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+V1.0.15
+- Wildcard topic pub / sub
+- Message replay on reconnect
+
 V1.0.8
 - History API fetch() to consume()